Normalize API versions fields across the project.
They should be uint16_t and it's not necessary enforce a sub_api field.
This patch also add some API checks in some places.

+++ src/lib/flow/include/sol-flow-parser.h
@@ -72,8 +72,8 @@ struct sol_flow_parser;
  */
 struct sol_flow_parser_client {
 #ifndef SOL_NO_API_VERSION
-#define SOL_FLOW_PARSER_CLIENT_API_VERSION (1UL)
-    unsigned long api_version; /**< @brief API version */
+#define SOL_FLOW_PARSER_CLIENT_API_VERSION (1)
+    uint16_t api_version; /**< @brief API version */
 #endif
     void *data; /**< @brief Client data */
